move new tests to dialog specific blocks
